Ensure both v1 and v2 SDF API endpoints consistently trigger and wait for index updates on change set creation
While the v1 endpoint was triggering the new change set message for
Edda, only the v2 endpoint was actually waiting for the index update to
happen.

The method used by v2 to do the trigger-and-wait is now in sdf-core
where it can be used by both v1 and v2 endpoints.
