+++ lib/sdf-server/src/service/v2/change_set.rs
@@ -27,7 +27,6 @@ use dal::{
     workspace_integrations::WorkspaceIntegration,
     workspace_snapshot::dependent_value_root::DependentValueRootError,
 };
-use futures_lite::StreamExt;
 use reqwest::Client;
 use sdf_core::{
     api_error::ApiError,
@@ -65,6 +64,8 @@ pub enum Error {
     ChangeSetApply(#[from] dal::ChangeSetApplyError),
     #[error("change set approval error: {0}")]
     ChangeSetApproval(#[from] dal::change_set::approval::ChangeSetApprovalError),
+    #[error("change set mvs error: {0}")]
+    ChangeSetMvs(#[from] sdf_core::change_set_mvs::ChangeSetMvsError),
     #[error("component error: {0}")]
     Component(#[from] dal::ComponentError),
     #[error("dal wrapper error: {0}")]
@@ -176,8 +177,6 @@ pub type ChangeSetAPIError = Error;
 
 type Result<T> = result::Result<T, Error>;
 
-const WATCH_INDEX_TIMEOUT: Duration = Duration::from_secs(30);
-
 #[derive(Serialize)]
 struct SlackMessage<'a> {
     text: &'a str,
@@ -255,46 +254,6 @@ pub fn change_set_routes(state: AppState) -> Router<AppState> {
             post(request_approval::request_approval),
         )
         .nest("/index", super::index::v2_change_set_routes())
-}
-
-#[instrument(
-    level = "info",
-    name = "sdf.change_set.create_index_for_new_change_set_and_watch",
-    skip_all,
-    fields(
-        si.edda_request.id = Empty
-    )
-)]
-pub async fn create_index_for_new_change_set_and_watch(
-    frigg: &frigg::FriggStore,
-    edda_client: &edda_client::EddaClient,
-    workspace_pk: WorkspacePk,
-    change_set_id: ChangeSetId,
-    base_change_set_id: ChangeSetId,
-    to_snapshot_address: WorkspaceSnapshotAddress,
-) -> Result<bool> {
-    let span = Span::current();
-    let mut watch = frigg
-        .watch_change_set_index(workspace_pk, change_set_id)
-        .await?;
-    let request_id = edda_client
-        .new_change_set(
-            workspace_pk,
-            change_set_id,
-            base_change_set_id,
-            to_snapshot_address,
-        )
-        .await?;
-    span.record("si.edda_request.id", request_id.to_string());
-
-    let timeout = WATCH_INDEX_TIMEOUT;
-    tokio::select! {
-        _ = tokio::time::sleep(timeout) => {
-            info!("timed out waiting for new change set index to be created");
-            Ok(false)
-        },
-        _ = watch.next() => Ok(true)
-    }
 }
 
 #[instrument(
